+++ src/Exception/MissingExtensionException.php
@@ -7,7 +7,7 @@ namespace Mews\PosBundle\Exception;
  */
 class MissingExtensionException extends \RuntimeException
 {
-    public function __construct($message = '', \Throwable $previous = null)
+    public function __construct(string $message = '', ?\Throwable $previous = null)
     {
         parent::__construct($message, 0, $previous);
     }