+++ src/Gateway/Builder/AbstractGatewayDefinitionBuilder.php
@@ -27,16 +27,6 @@ abstract class AbstractGatewayDefinitionBuilder implements GatewayDefinitionBuil
 
     protected function configureOptions(OptionsResolver $resolver): void
     {
-        $resolver->setDefaults([
-            'lang'      => PosInterface::LANG_TR,
-            'test_mode' => false,
-        ]);
-        $resolver->setAllowedValues('lang', [
-            PosInterface::LANG_TR,
-            PosInterface::LANG_EN,
-        ]);
-        $resolver->setAllowedTypes('test_mode', 'boolean');
-
         $resolver
             ->setRequired('gateway_class')
             ->setAllowedTypes('gateway_class', 'string');
@@ -57,18 +47,8 @@ abstract class AbstractGatewayDefinitionBuilder implements GatewayDefinitionBuil
         $this->setNestedOptions($resolver, 'credentials', function (OptionsResolver $subResolver): void {
             $subResolver->setRequired([
                 'merchant_id',
-                'payment_model',
-                'enc_key', // for 3D models only
             ]);
             $subResolver->setAllowedTypes('merchant_id', ['int', 'string']);
-            $subResolver->setAllowedTypes('enc_key', ['int', 'string']);
-            $subResolver->setAllowedValues('payment_model', [
-                PosInterface::MODEL_NON_SECURE,
-                PosInterface::MODEL_3D_SECURE,
-                PosInterface::MODEL_3D_PAY,
-                PosInterface::MODEL_3D_PAY_HOSTING,
-                PosInterface::MODEL_3D_HOST,
-            ]);
         });
 
         $this->setNestedOptions($resolver, 'gateway_configs', function (OptionsResolver $subResolver): void {
@@ -76,6 +56,8 @@ abstract class AbstractGatewayDefinitionBuilder implements GatewayDefinitionBuil
                 ->setAllowedTypes('test_mode', 'boolean');
             $subResolver->setDefault('disable_3d_hash_check', false)
                 ->setAllowedTypes('disable_3d_hash_check', 'boolean');
+            $subResolver->setDefault('lang', PosInterface::LANG_TR)
+                ->setAllowedValues('lang', [PosInterface::LANG_TR, PosInterface::LANG_EN]);
         });
     }