+++ src/Gateway/Builder/Param3DHostPosDefinitionBuilder.php
@@ -2,18 +2,14 @@
 
 namespace Mews\PosBundle\Gateway\Builder;
 
-use Mews\Pos\Gateways\EstPos;
-use Mews\Pos\Gateways\EstV3Pos;
+use Mews\Pos\Gateway\Param3DHostPos;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class EstV3PosDefinitionBuilder extends AbstractGatewayDefinitionBuilder
+class Param3DHostPosDefinitionBuilder extends AbstractGatewayDefinitionBuilder
 {
-    /**
-     * @inheritDoc
-     */
     public function supports(string $gatewayClass): bool
     {
-        return \in_array($gatewayClass, [EstPos::class, EstV3Pos::class], true);
+        return Param3DHostPos::class === $gatewayClass;
     }
 
     protected function getRequiredExtensions(): array
@@ -29,15 +25,23 @@ class EstV3PosDefinitionBuilder extends AbstractGatewayDefinitionBuilder
             $subResolver->setRequired([
                 'user_name',
                 'user_password',
+                'secret_key',
             ]);
             $subResolver->setAllowedTypes('user_name', ['int', 'string']);
             $subResolver->setAllowedTypes('user_password', ['int', 'string']);
+            $subResolver->setAllowedTypes('secret_key', ['int', 'string']);
+            $subResolver->setDefined('terminal_id')
+                ->setAllowedTypes('terminal_id', ['int', 'string']);
         });
-
     }
 
     protected function getRequiredEndpoints(): array
     {
-        return \array_merge(parent::getRequiredEndpoints(), ['gateway_3d']);
+        return \array_merge(parent::getRequiredEndpoints(), ['gateway_3d_host']);
+    }
+
+    protected function getOptionalEndpoints(): array
+    {
+        return [];
     }
 }