+++ src/Gateway/Builder/VakifKatilimPosDefinitionBuilder.php
@@ -2,7 +2,7 @@
 
 namespace Mews\PosBundle\Gateway\Builder;
 
-use Mews\Pos\Gateways\VakifKatilimPos;
+use Mews\Pos\Gateway\VakifKatilimPos;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class VakifKatilimPosDefinitionBuilder extends AbstractGatewayDefinitionBuilder
@@ -12,7 +12,7 @@ class VakifKatilimPosDefinitionBuilder extends AbstractGatewayDefinitionBuilder
      */
     public function supports(string $gatewayClass): bool
     {
-        return \in_array($gatewayClass, [VakifKatilimPos::class], true);
+        return VakifKatilimPos::class === $gatewayClass;
     }
 
     protected function getRequiredExtensions(): array
@@ -27,18 +27,15 @@ class VakifKatilimPosDefinitionBuilder extends AbstractGatewayDefinitionBuilder
         $this->setNestedOptions($resolver, 'credentials', function (OptionsResolver $subResolver): void {
             $subResolver->setRequired([
                 'terminal_id',
-                'user_name'
+                'user_name',
+                'secret_key',
             ]);
             $subResolver->setAllowedTypes('terminal_id', ['int', 'string']);
             $subResolver->setAllowedTypes('user_name', ['int', 'string']);
+            $subResolver->setAllowedTypes('secret_key', ['int', 'string']);
             $subResolver
                 ->setDefined('sub_merchant_id')
                 ->setAllowedTypes('sub_merchant_id', ['int', 'string']);
         });
-    }
-
-    protected function getRequiredEndpoints(): array
-    {
-        return \array_merge(parent::getRequiredEndpoints(), ['gateway_3d']);
     }
 }